MDStudio: OpenGL rendering refactor

+++ ion/renderer/IndexBuffer.h
@@ -18,21 +18,21 @@ namespace ion
 		class IndexBuffer
 		{
 		public:
-			void Add(u16 index) { mIndices.push_back(index); }
-			void Add(u16 face0, u16 face1, u16 face2)
+			void Add(u32 index) { mIndices.push_back(index); }
+			void Add(u32 face0, u32 face1, u32 face2)
 			{
 				mIndices.push_back(face0);
 				mIndices.push_back(face1);
 				mIndices.push_back(face2);
 			}
 
-			u16 Get(int index) const { return mIndices[index]; }
+			u32 Get(int index) const { return mIndices[index]; }
 
 			int GetSize() const { return mIndices.size(); }
-			const u16* GetAddress() const { return &mIndices[0]; }
+			const u32* GetAddress() const { return &mIndices[0]; }
 
 		private:
-			std::vector<u16> mIndices;
+			std::vector<u32> mIndices;
 		};
 	}
 }
